backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.0.3
openpyxl==3.1.2
firebase-admin==6.2.0
pydantic==2.4.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
